requirements.txt
apache-airflow 
apache-airflow-providers-apache-spark 
pyspark